+++ src/renderer/index.scss
@@ -21,3 +21,127 @@ td:empty::after,
 .place-held:empty::after {
   content: '\00a0';
 }
+// Global styles
+div#app {
+  width: 100vw;
+  height: 100vh;
+}
+
+.image-button-dark {
+  background: #3A3A3A;
+  border: #3A3A3A;
+  border-radius: 35px;
+}
+
+.image-button-dark:hover {
+  filter: invert(60%);
+}
+
+.image-button-dark:focus:not(.hover) {
+  border: none;
+  box-shadow: 0 0 0 4px #6c757d;
+  outline: none;
+}
+
+.image-button-dark:focus { // focus and hover
+  border: none;
+  box-shadow: 0 0 0 4px #808080;
+  outline: none;
+}
+
+.image-button {
+  background: white;
+  border: white;
+  border-radius: 35px;
+}
+
+.image-button:hover {
+  filter: invert(60%);
+}
+
+.image-button:focus:not(.hover) {
+  border: none;
+  box-shadow: 0 0 0 4px #6c757d;
+  outline: none;
+}
+
+.image-button:focus { // focus and hover
+  border: none;
+  box-shadow: 0 0 0 4px #808080;
+  outline: none;
+}
+
+::-webkit-scrollbar {
+  width: 0;
+  background: transparent;
+}
+
+// .alert-enter {
+//   opacity: 0;
+//   transform: scale(0.9);
+// }
+// .alert-enter-active {
+//   opacity: 1;
+//   transform: translateX(0);
+//   transition: opacity 3000ms, transform 3000ms;
+// }
+.slide-down-appear {
+  opacity: 0;
+  transform: translateY(-100%);
+}
+.slide-down-appear-active {
+  opacity: 1;
+  transform: translateY(0%);
+  transition: opacity 300ms, transform 300ms;
+}
+
+.slide-left-appear {
+  opacity: 0;
+  transform: translateX(100%);
+}
+.slide-left-appear-active {
+  opacity: 1;
+  transform: translateX(0%);
+  transition: opacity 300ms, transform 300ms;
+}
+
+.slide-right-appear {
+  opacity: 0;
+  transform: translateX(-100%);
+}
+.slide-right-appear-active {
+  opacity: 1;
+  transform: translateX(0%);
+  transition: opacity 300ms, transform 300ms;
+}
+.config-rec-set-appear {
+  opacity: 0;
+  transform: translateY(100%);
+}
+.config-rec-set-appear-active {
+  opacity: 1;
+  transform: translateY(0%);
+  transition: opacity 300ms, transform 300ms;
+}
+.config-rec-set-exit {
+  opacity: 1;
+  transform: translateY(0%);
+}
+.config-rec-set-exit-active {
+  opacity: 0;
+  transform: translateY(-100%);
+  transition: opacity 300ms, transform 300ms;
+}
+
+// Unless specified otherwise, all images are not draggable and not selectable
+* {
+  -webkit-user-drag: none;
+  user-select: none;
+}
+
+// https://stackoverflow.com/a/42225797
+td:empty::after,
+.list-group-item:empty::after,
+.place-held:empty::after {
+  content: '\00a0';
+}
